% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.R
\name{get_cor}
\alias{get_cor}
\title{Extract .cor file from xpose object}
\usage{
get_cor(xpdb)
}
\arguments{
\item{xpdb}{an object of "xpose_data" class}
}
\value{
cor file in the form of a tibble dataset.
}
\description{
Extract .cor file from xpose object
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
get_cor(x)
}
