% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_xpdb.R
\name{get_ext}
\alias{get_ext}
\title{Extract .ext file from xpose object}
\usage{
get_ext(xpdb)
}
\arguments{
\item{xpdb}{an object of "xpose_data" class}
}
\value{
ext file in the form of a tibble dataset.
}
\description{
Extract .ext file from xpose object
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
get_ext(x)
}
