% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.R
\name{get_lst}
\alias{get_lst}
\title{Extract .lst file from xpose object}
\usage{
get_lst(xpdb)
}
\arguments{
\item{xpdb}{an object of "xpose_data" class}
}
\value{
lst file in the form of a character vector.
}
\description{
Extract .lst file from xpose object
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
get_lst(x)
}
