% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_xpdb.R
\name{get_phi}
\alias{get_phi}
\title{Extract .phi file from xpose object}
\usage{
get_phi(xpdb)
}
\arguments{
\item{xpdb}{an object of "xpose_data" class}
}
\value{
phi file in the form of a tibble dataset.
}
\description{
Extract .phi file from xpose object
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
uncrtnty:::get_phi(x)
}
