% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockform.R
\name{infer_blockform}
\alias{infer_blockform}
\title{Infer Block Form from a Matrix}
\usage{
infer_blockform(x)
}
\arguments{
\item{x}{a symmetric matrix of numeric}
}
\value{
a vector of integer describing the blocks in the matrix.
}
\description{
Infer Block Form from a Matrix
}
\examples{
m <- matrix(c(1, 2, 0, 2, 3, 0, 0, 0, 4), nrow = 3)
infer_blockform(m)
}
