% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncrtnty.R
\name{is.symmetric}
\alias{is.symmetric}
\title{Symmetric matrix}
\usage{
is.symmetric(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
returns TRUE if the argument is a symmetric square matrix, or FALSE with a message otherwise
}
\description{
Symmetric matrix
}
\examples{
M <- matrix(c(11, 12, 13, 12, 22, 23, 13, 23, 33), nrow = 3)
is.symmetric(M) #is TRUE
M[1,2] <- 99
is.symmetric(M) #is FALSE with a message
}
