% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockform.R
\name{matrix_to_list}
\alias{matrix_to_list}
\title{Transpose a full matrix into a list of matrices}
\usage{
matrix_to_list(x, blockform = integer(0))
}
\arguments{
\item{x}{a symmetric matrix of numeric}

\item{blockform}{integer vector, explicitly defined by the user or parsed from a .lst file with `parse_lst()`. If empty (the default), will be infered with `infer_blockform()`.}
}
\value{
a list of matrices
}
\description{
Transpose a full matrix into a list of matrices
}
\examples{
m <- matrix(c(1,0.1,0,0,0.1,2,0,0,0,0,.3,0,0,0,0,.4), ncol = 4)
matrix_to_list(m)
}
