% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_nwpri.R
\name{nwpri_matrix}
\alias{nwpri_matrix}
\title{Set a block of matrix for the NWPRI routine in NONMEM}
\usage{
nwpri_matrix(x, blocktype)
}
\arguments{
\item{x}{a symmetric matrix of numeric, typically the covariance matrix of theta or the estimates of omega/sigma.}

\item{blocktype}{a character, either "THETAP", "OMEGAPD" or "SIGMAPD".}
}
\value{
a single character value to be pasted into the NONMEM code
}
\description{
Set a block of matrix for the NWPRI routine in NONMEM
}
\examples{
x <- matrix(c(1, 0.12, 0.13, 0.12, 22, 0.23, 0.13, 0.23, 333), nrow = 3, ncol = 3)
nwpri_matrix(x = x, blocktype = "THETAPV")
nwpri_matrix(x = x, blocktype = "OMEGAP")
}
