% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_nwpri.R
\name{nwpri_value}
\alias{nwpri_value}
\title{Set a block of values for the NWPRI routine in NONMEM}
\usage{
nwpri_value(x, blocktype)
}
\arguments{
\item{x}{a (named) vector of numeric, typically the estimates of theta or the degree of freedom associated to omega/sigma.}

\item{blocktype}{a character, either "THETAP", "OMEGAPD" or "SIGMAPD".}
}
\value{
a single character value to be pasted into the NONMEM code
}
\description{
Set a block of values for the NWPRI routine in NONMEM
}
\examples{
nwpri_value(x = c(CL = .123, 456, V = 789), blocktype = "THETAP")
nwpri_value(x = c(123, 45, 67), blocktype = "OMEGAPD")
}
