% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockform.R
\name{parse_blockform_from_lst}
\alias{parse_blockform_from_lst}
\title{Parse Block Form from a .lst NONMEM output}
\usage{
parse_blockform_from_lst(lst)
}
\arguments{
\item{lst}{lst file in the form of a character vector.}
}
\value{
a vector of integer describing the blocks in the matrix.
}
\description{
Parse Block Form from a .lst NONMEM output
}
\examples{
lst3a <- readLines(system.file("nm", "run003a.lst", package = "uncrtnty"))
parse_blockform_from_lst(lst = lst3a)
}
