% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nmoutput.R
\name{parse_cor}
\alias{parse_cor}
\title{Parse a correlation table to useful matrices}
\usage{
parse_cor(tab, matrix = c("THETA", "OMEGA", "SIGMA"))
}
\arguments{
\item{tab}{a correlation table from NONMEM}

\item{matrix}{whether the `THETA`, `OMEGA` or `SIGMA` matrix should be parsed ?}
}
\value{
a matrix of numeric. If matrix is `THETA`, the variance-covariance matrix of estimation of `THETA`. If matrix is `OMEGA`/`SIGMA`, a matrix of the standard errors of every elements of the `OMEGA`/`SIGMA` matrices.
}
\description{
Parse a correlation table to useful matrices
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
cor <- get_cor(x)
parse_cor(cor, matrix = "THETA")
parse_cor(cor, matrix = "OMEGA")
}
