% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nmoutput.R
\name{parse_cov}
\alias{parse_cov}
\title{Parse a covariance table to useful matrices}
\usage{
parse_cov(tab)
}
\arguments{
\item{tab}{a covariance table from NONMEM}
}
\value{
a list with the variance-covariance matrix of estimation of `THETA`, and the matrices of the standard errors of every elements of the `OMEGA`/`SIGMA` matrices.
}
\description{
Parse a covariance table to useful matrices
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
cov <- get_cov(x)
parse_cov(cov)
}
