% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nmoutput.R
\name{parse_ext}
\alias{parse_ext}
\title{Parse a ext table to a list with final estimates}
\usage{
parse_ext(tab)
}
\arguments{
\item{tab}{a ext table from NONMEM}
}
\value{
a list, with a vector of `THETA` and `OMEGA`/`SIGMA` matrices
}
\description{
Parse a ext table to a list with final estimates
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
ext <- get_ext(x)
parse_ext(ext)
}
