% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nmoutput.R
\name{parse_lst}
\alias{parse_lst}
\title{Parse a lst file to a list of relevant information}
\usage{
parse_lst(lst)
}
\arguments{
\item{lst}{lst file in the form of a character vector.}
}
\value{
a list with the number of individuals, number of observations, and the "blockforms" of omega and sigma matrices
}
\description{
Parse a lst file to a list of relevant information
}
\examples{
lst3a <- readLines(system.file("nm", "run003a.lst", package = "uncrtnty"))
parse_lst(lst = lst3a)
lst001 <- readLines(system.file("xposerun", "run001.lst", package = "uncrtnty"))
parse_lst(lst = lst001)
}
