% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nmoutput.R
\name{parse_phi}
\alias{parse_phi}
\title{Parse a phi table to useful individual etas and covariance matrix}
\usage{
parse_phi(tab)
}
\arguments{
\item{tab}{a phi table from NONMEM}
}
\value{
a list (length = number of individual), with the vector of estimated etas and the associated variance-covariance matrix of estimation
}
\description{
Parse a phi table to useful individual etas and covariance matrix
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
phi <- get_phi(x)
parse_phi(head(phi, 3)) #first 3 individuals for the example
}
