% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_litt.R
\name{reporttype}
\alias{reporttype}
\alias{as_cv}
\alias{as_cve1}
\alias{as_rse}
\alias{as_se}
\alias{as_ci95}
\alias{as_ci90}
\title{Attribute a "report type"}
\usage{
as_cv(x)

as_cve1(x)

as_rse(x)

as_se(x)

as_ci95(x)

as_ci90(x)
}
\arguments{
\item{x}{list of values}
}
\description{
Attribute a "report type"
}
