% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{u_example}
\alias{u_example}
\title{An example of uncrtnty-list object}
\format{
An object of class \code{uncrtnty} of length 7.
}
\usage{
u_example
}
\description{
An example of uncrtnty-list object
}
\keyword{datasets}
