% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_litt.R
\name{u_from_litt}
\alias{u_from_litt}
\title{Build uncrtnty-list from literature}
\usage{
u_from_litt(
  model = "Prior model",
  nid = Inf,
  nobs = Inf,
  th_est,
  th_unc,
  om_est,
  om_unc,
  si_est,
  si_unc
)
}
\arguments{
\item{model}{a character, model name}

\item{nid, nobs}{number of individuals (observations) used to build the model. Will be used as a maximum for `OMEGA` (`SIGMA`) degrees of freedom.}

\item{th_est}{a vector of numeric, point estimates of `THETA`}

\item{th_unc}{a full variance-covariance matrix of estimation of `THETA`, alternatively a list wrapped inside `as_rse()`, `as_se()`, `as_ci95()` or `as_ci90()` (see details)}

\item{om_est}{a list of variances or variance-covariance matrices, alternatively a list wrapped inside `as_cv()` or `as_cve1()`}

\item{om_unc}{a list wrapped inside `as_rse()`, `as_ci95()` or `as_ci90()`}

\item{si_est}{a list of variances or variance-covariance matrices, alternatively a list wrapped inside `as_cv()` or `as_cve1()`}

\item{si_unc}{a list wrapped inside `as_rse()`, `as_ci95()` or `as_ci90()`}
}
\value{
an uncrtnty-list object
}
\description{
Build uncrtnty-list from literature
}
\examples{
u_from_litt(
model = "Test", nid = 250, nobs = 12345, #initial model
th_est = c(1,22,333,0.444),
th_unc = as_rse(list(0.2, 0.25, 0.3, .80)), #RSE on THETA are 20\%, 25\%, 30\% and 80\%
om_est = list(0.25, 0.33, 0.88), #omega reported as variances, i.e. OMEGA2
om_unc = as_ci95(list(c(0.2, 0.3), c(0.16, 0.50), c(0.8, 0.96))), #CI from bootstrap for instance
si_est = as_cv(list(0.2)), #sigma reported as a CV\%, for a proportional error for instance
si_unc = as_rse(list(0.05)) #5\% RSE
)

}
