% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_xpdb.R
\name{u_from_xpdb}
\alias{u_from_xpdb}
\title{Build uncrtnty-list from xpdb object}
\usage{
u_from_xpdb(xpdb)
}
\arguments{
\item{xpdb}{an object of "xpose_data" class}
}
\value{
an uncrtnty-list object
}
\description{
Build uncrtnty-list from xpdb object
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
u <- u_from_xpdb(x)
class(u)
}
