% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_simpar.R
\name{u_to_simpar}
\alias{u_to_simpar}
\title{Turn uncrtnty-list into list of arguments for simpar}
\usage{
u_to_simpar(u, nsim, ...)
}
\arguments{
\item{u}{an uncrtnty-list object}

\item{nsim}{scalar numeric specifying the number of sets to attempt}

\item{...}{additional named arguments to be passed to `simpar::simpar()` (i.e. "digits", "min", "max").}
}
\value{
a list of arguments to be called in `simpar::simpar()`.
}
\description{
Turn uncrtnty-list into list of arguments for simpar
}
\details{
The arguments "theta", "covar", "omega", "sigma", "odf" and "sdf" are filled from the information contained in the uncrtnty-list object, unless they are explicitly specified in `...` .
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
u <- u_from_xpdb(x)
args <- u_to_simpar(u, n = 10, sdf = 1000, digits = 5, max = Inf)
#remotes::install_github("metrumresearchgroup/simpar")
#do.call(simpar::simpar, args)

}
