% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncrtnty.R
\name{uncrtnty}
\alias{uncrtnty}
\title{Build an `uncrtnty` list object}
\usage{
uncrtnty(
  model = character(),
  nid = integer(),
  nobs = integer(),
  th_est = double(),
  th_unc = matrix(),
  om_est = list(),
  om_unc = integer(),
  si_est = list(),
  si_unc = integer()
)
}
\arguments{
\item{model}{character, simple description of the model}

\item{nid}{integer, the number of subjects in the orginal analysis}

\item{nobs}{integer, the number of observations in the original analysis}

\item{th_est}{double vector, estimations of theta}

\item{th_unc}{double matrix, uncertainty of theta, under the form of a covariance matrix}

\item{om_est}{list of double matrices, estimations of omega}

\item{om_unc}{integer vector, uncertainty of omega, under the form of degrees of freedom}

\item{si_est}{list of double matrices, estimations of sigma}

\item{si_unc}{integer vector, uncertainty of sigma, under the form of degrees of freedom}
}
\value{
an `uncrtnty` list object.
}
\description{
Build an `uncrtnty` list object
}
\details{
This function is not meant to be used by the user, but by more user-friendly functions such as `u_from_xpdb()` for instance.
}
\examples{
u_example
}
