% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_simpar.R
\name{xpdb_to_simpar}
\alias{xpdb_to_simpar}
\title{Get uncertainty data from xpose object}
\usage{
xpdb_to_simpar(xpdb)
}
\arguments{
\item{xpdb}{an object of "xpose_data" class}
}
\value{
a list of estimations and their uncertainties, suited to be passed to `simpar::simpar()`
}
\description{
Get uncertainty data from xpose object
}
\examples{
x <- readRDS(system.file("xposerun", "xpdb_ex_pk.rds", package = "uncrtnty"))
xpdb_to_simpar(x)
}
